Reorganize module structure for better type organization
- Create dedicated `chunk` module with Chunk, MarkdownChunk, RetrievedChunk, Metric
- Create dedicated `embedding` module with EmbeddingProvider, EmbeddingOpenAI
- Move Document and MarkdownDocument definitions to `document` module
- Move BaseChunker to `_chunker` module alongside RaghildaMarkdownChunker
- Rename Store to BaseStore for consistency with BaseChunker
- Simplify `_types` module to only contain protocol definitions
- Update `types` module to export only protocol types (ChunkLike, DocumentLike, etc.)
- Update all cross-module imports to use public modules instead of internal ones
- Update tests to use new import paths

New module structure:
- raghilda.chunk: Chunk, MarkdownChunk, RetrievedChunk, Metric
- raghilda.document: Document, MarkdownDocument
- raghilda.embedding: EmbeddingProvider, EmbeddingOpenAI
- raghilda.store: BaseStore, DuckDBStore, OpenAIStore
- raghilda.chunker: BaseChunker, RaghildaMarkdownChunker, MarkdownChunk
- raghilda.types: ChunkLike, DocumentLike, ChunkerLike, IntoChunk, IntoDocument
