+++ src/raghilda/document.py
@@ -1,6 +1,46 @@
 from dataclasses import dataclass, field
 from typing import Optional, Union
-from ._types import Chunk, Document, DocumentLike, IntoDocument
+import uuid
+
+from .types import DocumentLike, IntoDocument
+from .chunk import Chunk
+
+__all__ = ["Document", "MarkdownDocument"]
+
+
+def _generate_doc_id() -> str:
+    return f"doc_{uuid.uuid4().hex}"
+
+
+@dataclass
+class Document:
+    """Base document type for raghilda."""
+
+    content: str
+    id: str = field(default_factory=_generate_doc_id)
+    chunks: Optional[list[Chunk]] = None
+
+    @classmethod
+    def from_any(cls, doc: Union[DocumentLike, IntoDocument]) -> "Document":
+        """Convert any document-like or IntoDocument object to a raghilda Document."""
+        if isinstance(doc, IntoDocument):
+            if not callable(doc.to_document):
+                raise TypeError(
+                    f"{type(doc).__name__}.to_document must be a method, not {type(doc.to_document).__name__}"
+                )
+            result = doc.to_document()
+            if not isinstance(result, Document):
+                raise TypeError(
+                    f"{type(doc).__name__}.to_document() must return a Document, got {type(result).__name__}"
+                )
+            return result
+        elif isinstance(doc, DocumentLike):
+            chunks = None
+            if doc.chunks is not None:
+                chunks = [Chunk.from_any(c) for c in doc.chunks]
+            doc_id = getattr(doc, "id", None) or _generate_doc_id()
+            return cls(content=doc.content, id=doc_id, chunks=chunks)
+        raise TypeError(f"Cannot convert {type(doc).__name__} to Document")
 
 
 @dataclass
@@ -19,14 +59,3 @@ class MarkdownDocument(Document):
             chunks=base.chunks,
             origin=getattr(doc, "origin", origin),
         )
-
-
-@dataclass
-class Metric:
-    name: str
-    value: float
-
-
-@dataclass
-class RetrievedChunk(Chunk):
-    metrics: list[Metric] = field(default_factory=list)
