+++ src/raghilda/embedding.py
@@ -1,0 +1,3 @@
+from ._embedding import EmbeddingProvider, EmbeddingOpenAI
+
+__all__ = ["EmbeddingProvider", "EmbeddingOpenAI"]
