+++ src/raghilda/store.py
@@ -1,4 +1,4 @@
-from ._store import Store, DuckDBStore
+from ._store import BaseStore, DuckDBStore
 from ._openai_store import OpenAIStore
 
-__all__ = ["Store", "DuckDBStore", "OpenAIStore"]
+__all__ = ["BaseStore", "DuckDBStore", "OpenAIStore"]
