+++ src/raghilda/types.py
@@ -1,29 +1,18 @@
-"""Public types for raghilda.
+"""Protocol types for raghilda.
 
-This module exports the core types and protocols for working with chunks,
-documents, and chunkers in raghilda.
+This module exports the protocol types for type checking compatibility
+with chunks, documents, and chunkers.
 """
 
 from ._types import (
-    # Concrete types
-    Chunk,
-    Document,
-    BaseChunker,
-    # Protocols for accepting compatible types
     ChunkLike,
     DocumentLike,
     ChunkerLike,
-    # Protocols for custom conversion
     IntoChunk,
     IntoDocument,
 )
 
 __all__ = [
-    # Concrete types
-    "Chunk",
-    "Document",
-    "BaseChunker",
-    # Protocols
     "ChunkLike",
     "DocumentLike",
     "ChunkerLike",
