+++ tests/test_chonkie_compat.py
@@ -6,9 +6,10 @@ import pytest
 chonkie = pytest.importorskip("chonkie")
 
 from chonkie.types import Chunk as ChonkieChunk, Document as ChonkieDocument  # noqa: E402
-from raghilda._types import Chunk, Document, ChunkLike, DocumentLike  # noqa: E402
-from raghilda._store import DuckDBStore  # noqa: E402
-from raghilda.document import MarkdownDocument  # noqa: E402
+from raghilda.chunk import Chunk  # noqa: E402
+from raghilda.document import Document, MarkdownDocument  # noqa: E402
+from raghilda.types import ChunkLike, DocumentLike  # noqa: E402
+from raghilda.store import DuckDBStore  # noqa: E402
 
 
 class TestChonkieChunkCompatibility:
